Cambios sustanciales a Anexo y su Clase
