+++ Anexo.py
@@ -5,21 +5,27 @@ from AtributosAnexo import AtributosAnexo
 
 @dataclass
 class Anexo:
-    fechaVersion: datetime 
+    texto: str
     metadatoAnexo: MetadatoAnexo
-    atributos: AtributosAnexo
+    idParte: str
+    fechaVersion: datetime
+    derogado: str
+    transitorio: str
 
-    def __init__(self, fechaVersion=None, metadatoAnexo=None, atributos=None) -> None:  # Inician vacios por si hay algún problema, puede iniciarse con o sin datos
+    def __init__(self, texto=None, metadatoAnexo=None, idParte=None, fechaVersion=None, derogado=None, transitorio=None):  # Inician vacios por si hay algún problema, puede iniciarse con o sin datos
         # Validación de datos
-        self.setFechaVersion(fechaVersion)
+        self.setTexto(texto)
         self.setMetadatoAnexo(metadatoAnexo)
-        self.setAtributos(atributos)
+        self.setIdParte(idParte)
+        self.setFechaVerision(fechaVersion)
+        self.setDerogado(derogado)
+        self.setTransitorio(transitorio)
 
-    def setFechaVersion(self, fechaVersion: datetime) -> None:
-        if fechaVersion is None:
-            self.fechaVersion = "Fecha no encontrada"
+    def setTexto(self, texto: str) -> None:
+        if texto is None:
+            self.texto = "Texto no encontrado"
         else:
-            self.fechaVersion = fechaVersion
+            self.texto = texto
 
     def setMetadatoAnexo(self, metadatoAnexo: MetadatoAnexo) -> None:
         if metadatoAnexo is None:
@@ -27,17 +33,45 @@ class Anexo:
         else:
             self.metadatoAnexo = metadatoAnexo
 
-    def setAtributos(self, atributos: AtributosAnexo) -> None:
-        if atributos is None:
-            self.atributos = AtributosAnexo()
+    def setIdParte(self, idParte: str) -> None:
+        if idParte is None:
+            self.idParte = 0  # Error por defecto
         else:
-            self.atributos = atributos
+            self.idParte = idParte
 
-    def getFechaVersion(self) -> datetime:
-        return self.fechaVersion
+    def setFechaVerision(self, fechaVersion: datetime) -> None:
+        if fechaVersion is None:
+            self.fechaVersion = datetime(1800,1,1)
+        else:
+            self.fechaVersion = fechaVersion
+
+    def setDerogado(self, derogado: str) -> None:
+        if derogado is None:
+            self.derogado = "desconocido"
+        else:
+            self.derogado = derogado
+
+    def setTransitorio(self, transitorio: str) -> None:
+        if transitorio is None:
+            self.transitorio = "desconocido"
+        else:
+            self.transitorio = transitorio
+
+    def getTexto(self) -> str:
+        return self.texto
 
     def getMetadatoAnexo(self) -> MetadatoAnexo:
         return self.metadatoAnexo
 
-    def getAtributos(self) -> AtributosAnexo:
-        return self.atributos
+    def getIdParte(self) -> str:
+        return self.idParte
+
+    def getFechaVersion(self) -> datetime:
+        return self.fechaVersion
+
+    def getDerogado(self) -> str:
+        return self.derogado
+
+    def getTransitorio(self) -> str:
+        return self.transitorio
+
